+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Fixed integer C/C++ division, forced released by CRAN (#4185). 
 
+* `group_indices()` now ignores empty groups by default for `data.frame`, which is
+  consistent with the default of `group_by()` (@yutannihilation, #4208). 
+
 # dplyr 0.8.0 (2019-02-14)
 
 ## Breaking changes