+++ NEWS.md
@@ -1,13 +1,15 @@
 # dplyr 0.8.0.9000
 
+* Fixed handling of bare formulas in colwise verbs (#4183).
+
 * Fixed performance of `n_distint()` (#4202). 
+
+* `group_indices()` now ignores empty groups by default for `data.frame`, which is
+  consistent with the default of `group_by()` (@yutannihilation, #4208). 
 
 # dplyr 0.8.0.1 (2019-02-15)
 
 * Fixed integer C/C++ division, forced released by CRAN (#4185). 
-
-* `group_indices()` now ignores empty groups by default for `data.frame`, which is
-  consistent with the default of `group_by()` (@yutannihilation, #4208). 
 
 # dplyr 0.8.0 (2019-02-14)